pyproject.toml
[tool.poetry]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
drf-yasg = "^1.21.8"
django-filter = "^24.3"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
